requirements.txt
torch==2.7.1
torchvision==0.22.1
pillow==11.3.0
ultralytics==8.3.170
matplotlib==3.10.3
numpy==1.26.4
seaborn==0.13.2
scikit-learn==1.7.1
tqdm==4.67.1
pandas==2.3.1
pycocotools==2.0.10
opencv-python==4.11.0.86
onnx==1.17.0
onnx2tf==1.28.2
tensorflow==2.20.0
groundingdino-py
paho-mqtt
redis
Flask
tflite-runtime